SCP-1687: ChainServer: refactoring
