+++ plutus-pab/src/Plutus/PAB/PABLogMsg.hs
@@ -10,7 +10,7 @@
 module Plutus.PAB.PABLogMsg(
     PABLogMsg(..),
     ContractExeLogMsg(..),
-    ChainIndexServerMsg(..),
+    ChainIndexServerMsg,
     SigningProcessMsg,
     AppMsg(..)
     ) where
@@ -20,13 +20,13 @@ import qualified Data.Aeson.Encode.Pretty           as JSON
 import qualified Data.ByteString.Lazy.Char8         as BSL8
 import           Data.String                        (IsString (..))
 import           Data.Text                          (Text)
-import           Data.Text.Prettyprint.Doc          (Pretty (..), colon, hang, parens, viaShow, vsep, (<+>))
+import           Data.Text.Prettyprint.Doc          (Pretty (..), colon, hang, viaShow, vsep, (<+>))
 import           Data.Time.Units                    (Second)
 import           GHC.Generics                       (Generic)
 
 import           Cardano.BM.Data.Tracer             (ToObject (..), TracingVerbosity (..))
 import           Cardano.BM.Data.Tracer.Extras      (Tagged (..), mkObjectStr)
-import           Cardano.Node.Types                 (FollowerID)
+import           Cardano.ChainIndex.Types           (ChainIndexServerMsg)
 import           Cardano.SigningProcess.Types       (SigningProcessMsg)
 import           Language.Plutus.Contract.State     (ContractRequest)
 import           Ledger.Tx                          (Tx)
@@ -56,40 +56,6 @@ data AppMsg =
     deriving stock (Show, Generic)
     deriving anyclass (ToJSON, FromJSON)
 
--- | Messages from the ChainIndex Server
-data ChainIndexServerMsg =
-    -- | Obtaining a new follower
-    ObtainingFollowerID
-    -- | Obtained a new follower 'FollowerID'
-    | ObtainedFollowerID FollowerID
-    -- | Updating the chain index with 'FollowerID'
-    | UpdatingChainIndex FollowerID
-    -- | Requesting new blocks from the node
-    | AskingNodeForNewBlocks
-    -- | Requesting the current slot from the node
-    | AskingNodeForCurrentSlot
-    -- | Starting a node client thread
-    | StartingNodeClientThread
-    -- | Starting ChainIndex service
-    | StartingChainIndex
-        Int    -- ^ Port number
-      -- | Received transaction
-    | ReceivedBlocksTxns
-        Int    -- ^ Blocks
-        Int    -- ^ Transactions
-    deriving stock (Show, Generic)
-    deriving anyclass (ToJSON, FromJSON)
-
-instance Pretty ChainIndexServerMsg where
-    pretty = \case
-        ObtainingFollowerID -> "Obtaining follower ID"
-        ObtainedFollowerID i -> "Obtained follower ID:" <+> pretty i
-        UpdatingChainIndex i -> "Updating chain index with follower ID" <+> pretty i
-        ReceivedBlocksTxns blocks txns -> "Received" <+> pretty blocks <+> "blocks" <+> parens (pretty txns <+> "transactions")
-        AskingNodeForNewBlocks -> "Asking the node for new blocks"
-        AskingNodeForCurrentSlot -> "Asking the node for the current slot"
-        StartingNodeClientThread -> "Starting node client thread"
-        StartingChainIndex port -> "Starting chain index on port: " <> pretty port
 
 instance Pretty AppMsg where
     pretty = \case
@@ -235,16 +201,6 @@ instance ToObject PABLogMsg where
         SSigningProcessMsg m   -> toObject v m
 
 
-instance ToObject ChainIndexServerMsg where
-    toObject _ = \case
-      ObtainingFollowerID      -> mkObjectStr "obtaining FollowerID" ()
-      ObtainedFollowerID fID   -> mkObjectStr "obtained FollowerID" (Tagged @"followerID" fID)
-      UpdatingChainIndex fID   -> mkObjectStr "updating chainIndex with FollowerID" (Tagged @"followerID" fID)
-      ReceivedBlocksTxns x y   -> mkObjectStr "received block transactions" (Tagged @"blocks" x, Tagged @"transactions" y)
-      AskingNodeForNewBlocks   -> mkObjectStr "asking for new blocks" ()
-      AskingNodeForCurrentSlot -> mkObjectStr "asking node for current slot" ()
-      StartingNodeClientThread -> mkObjectStr "starting node client thread" ()
-      StartingChainIndex p     -> mkObjectStr "starting chain index" (Tagged @"port" p)
 
 
 instance ToObject ContractExeLogMsg where
